refactor: remove duplicated functions and components
