+++ src/App.css
@@ -1,9 +1,3 @@
-.memo-app {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
 .memo-panel {
   margin: 1% auto;
   width: 300px;
